api-server/pyproject.toml
[tool.poetry]
name = "ticketcrm"
version = "0.1.0"
description = ""
authors = ["Andrea Bellomia <119261787+AndreaBellomia@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.3"
django = "^5.0.3"
djangorestframework = "^3.14.0"
markdown = "^3.5.2"
django-filter = "^24.1"
django-debug-toolbar = "^4.3.0"
psycopg2 = "^2.9.9"
django-rest-knox = "^4.2.0"
django-cors-headers = "^4.3.1"
drf-yasg = "^1.21.7"
gunicorn = "^22.0.0"
environs = "^11.0.0"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py311']
preview = true
skip-string-normalization = true
skip-magic-trailing-comma = true